test fixtures for centralizing of tests
