+++ core/build.gradle.kts
@@ -10,6 +10,7 @@ plugins {
     id("maven-publish")
     id("org.jreleaser") version "1.5.1"
     id("signing")
+    id("java-test-fixtures")
 }
 
 group = "de.tschuehly"
@@ -30,7 +31,10 @@ dependencies {
     implementation("com.fasterxml.jackson.module:jackson-module-kotlin")
     implementation("org.jetbrains.kotlin:kotlin-reflect")
     implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8")
-    testImplementation("org.springframework.boot:spring-boot-starter-test")
+    testFixturesImplementation("org.springframework.boot:spring-boot-starter-test")
+    testFixturesImplementation("org.springframework.boot:spring-boot-starter-web")
+    testFixturesImplementation("org.springframework.boot:spring-boot-devtools")
+
 }
 
 tasks.withType<KotlinCompile> {
