+++ examples/jte-example/build.gradle.kts
@@ -24,7 +24,10 @@ dependencies {
 
     implementation("org.webjars.npm:htmx.org:1.9.2")
     implementation("org.webjars:webjars-locator-core:0.53")
+
     testImplementation("org.springframework.boot:spring-boot-starter-test")
+    testImplementation("org.springframework.boot:spring-boot-devtools")
+    testImplementation(testFixtures("de.tschuehly:spring-view-component-core:0.7.3-SNAPSHOT"))
 }
 
 tasks.withType<Test> {
