+++ examples/kte-example/build.gradle.kts
@@ -31,7 +31,10 @@ dependencies {
     implementation("org.springframework.boot:spring-boot-devtools")
 
     implementation("org.jetbrains.kotlin:kotlin-reflect")
+
     testImplementation("org.springframework.boot:spring-boot-starter-test")
+    testImplementation("org.springframework.boot:spring-boot-devtools")
+    testImplementation(testFixtures("de.tschuehly:spring-view-component-core:0.7.3-SNAPSHOT"))
 }
 
 tasks.withType<KotlinCompile> {
